firestore/requirements.txt
google-cloud-firestore==2.5.3